[KFKESQ-11] Started refactoring of the consumer DSL
