+++ kafkaesque-core/src/main/java/in/rcard/kafkaesque/producer/AfterAllAssertions.java
@@ -1,0 +1,56 @@
+package in.rcard.kafkaesque.producer;
+
+import in.rcard.kafkaesque.producer.KafkaesqueProducer.Record;
+import java.time.Duration;
+import java.util.List;
+import java.util.function.Consumer;
+import org.apache.kafka.clients.producer.ProducerRecord;
+import org.awaitility.Awaitility;
+import org.awaitility.core.ConditionTimeoutException;
+
+class AfterAllAssertions<Key, Value> {
+  
+  private final KafkaesqueProducer<Key, Value> producer;
+  private final List<Record<Key, Value>> records;
+  private final Duration waitForConsumerDuration;
+  
+  AfterAllAssertions(
+      KafkaesqueProducer<Key, Value> producer,
+      List<Record<Key, Value>> records,
+      Duration waitForConsumerDuration) {
+    this.producer = producer;
+    this.records = records;
+    this.waitForConsumerDuration = waitForConsumerDuration;
+  }
+  
+  /**
+   * Asserts that some conditions hold on the whole list of sent message.<br> For example:
+   *
+   * <pre>
+   *   producer.asserting(records ->
+   *     assertThat(records).hasSize(2)
+   *   );
+   * </pre>
+   *
+   * @param messagesConsumer The conditions that must hold on the whole list of messages
+   */
+  void asserting(Consumer<List<ProducerRecord<Key, Value>>> messagesConsumer) {
+    final List<ProducerRecord<Key, Value>> producerRecords = producer.sendRecords(records);
+    consume(messagesConsumer, producerRecords);
+  }
+  
+  private void consume(
+      Consumer<List<ProducerRecord<Key, Value>>> messagesConsumer,
+      List<ProducerRecord<Key, Value>> records) {
+    try {
+      Awaitility.await()
+          .atMost(waitForConsumerDuration)
+          .untilAsserted(() -> messagesConsumer.accept(records));
+    } catch (ConditionTimeoutException ex) {
+      throw new AssertionError(
+          String.format(
+              "The consuming of the list of messages %s takes more than %d milliseconds",
+              records, waitForConsumerDuration.toMillis()));
+    }
+  }
+}
